fixup: Rename the undocumented GASNET_ALLOW_OPTIMIZED_DEBUG override
The e83f4dc change causes compilation failures in --enable-debug
mode on compilers that optimize by default (eg Pathscale ekopath)
when building delay.o for {udp,mpi}-conduit.

The fix is to propagate the same temporary workaroud used in gasnet.h
to the AMX equivalents in ammpi.h and amudp_const.h.

This commit should NOT be merged to develop (it *should* conflict).