gemini/aries: align immediate buffer in gpd
This commit makes two changes which together ensure that the 128-byte
"immediate buffer" in the gasnetc_post_descriptor_t (aka "gpd") is
64-byte aligned for maximum performance of the FMA operations which
use it as a bounce buffer.

1) The logic for allocating and initializing the gpd pool now uses a
64-byte aligned size rather than the actual size of the gdb.  Note,
however, that on Aries the size was *already* a multiple of 64 and on
Gemini is was just 16-bytes short.  So, there is NOT any increase in
memory usage on Aries and a default increase of just 8K on Gemini.

2) The union containing the immediate buffer is moved to be first in
the gpd structure, and thus benefit from the alignment from item #1.

Resolves bug 3358: Aries performance anomaly for 128-byte Gets