cygwin/mmap: re-re-fix bug3370
With this commit, mmap() support for Cygwin is once again disabled.
In fact, it is no longer possible to forcibly enable it.

The support for PSHM by default remains unchanged.  What has changed
is that gasnet_mmap.c now has a special case for Cygwin with PSHM
enabled over POSIX or FILE, in which the necessary mmap() calls are
allowed, even though general support for mmap() is not.