pshm: prevent one source of shared memory leaks
As detailed in bug 3373, certain errors in gasneti_attach() could
cause a race between gasneti_fatalerror()'s process-killing behavior
and the call to gasneti_cleanup_shm() in the killed processes.  This
was seen to result in a leak of shared memory objects.

By adding a supernode-scoped barrier after gasneti_cleanup_shm() we
ensure that the shared memory objects are "unlinked" prior to allowing
gasneti_segmentAttach() to return.