Add GASNETT_NORETURN to test.h FATALERR()
Manually specialize the definition of _test_doErrMsg into fatal and non-fatal
versions, so the fatal version called by FATALERR can get a
GASNETT_NORETURN attribute. This improves compiler analysis around
calls and helps suppress maybe-uninitialized warnings (such as the one in
testnbr, which no longer needs an explicit abort).

The only change to the _test_doErrMsg body is the removal of
the _test_fatalmsg global variable, which is now obsolete.