Factor conduit-independent AM registration code
This commit factors the conduit-independent portions of AM
registration, in paticular the logic which deals with assignment for
"don't care" entries and detection of duplicates.  The code to
actually "register" remains in the conduit, even though it is a
simple assignment in most cases.

Rewrite of EX commit ac1342817c845e6e9b7d6732d6ea0b57c00236fc