asm/atomics: cleanup support for PGI on PPC64el
This commit moves preprocesor logic for dealing with bugs in the PGI
compilers for PPC64el from gasnet_atomic_bits.h to gasnet_asm.h.
This centralizes the handling of such version based (as opposed to
configure probed) definitions.

Additionally, this changes gasnet_asm.h to treat the PPC version of
the PGI compilers separately from the X86(-64) version.