ibv,mxm,portals4: Resolve bug 2079
This commit removes the automatic disable of alloc() in ibv, mxm and
portals4 conduits (the subject of bug 2079).  It has been determined
that it is not reasonable to continue support for PGI compilers old
enough to display this bug, and we have been unable to reproduce the
problem on any compiler new enough to compile ibv-conduit at all.

The #if GASNETI_USE_ALLOCA remains as a guard for all uses of
alloca(), to allow it to be disabled again in the future if this
is necessary (perhaps just to debug a problem).