ofi: Merge the AM and RDMA transmit CQs into one
By merging the completion queues for outgoing operations into one,
the number of function calls in GASNet_AMPoll() is reduced. There
is also less duplicate code.

A task for the future is to determine if the TX and RX completion
queues can be merged as well, leaving only one CQ.