pshm: avoid write-back of mmaped file on FreeBSD
This commit ensures we pass MAP_NOSYNC to mmap() on FreeBSD when using
FILE or POSIX as the PSHM implementation, which tells the kernel not
to write dirty pages of the mapping to the backing file unless
necessary (paging due to memory pressure).

There is currently no other OS that provides an equivalent flag.