ofi: Enable FI_MR_BASIC memory registration mode
This commit wraps the fi_write and fi_read functions into a couple
macros that hide the details of determining what the addresses and
memory keys need to be. The intention is to be able to use a compile
time constant to compile away the branches required if you know what
memory mode that you support.