Add ofi-provider detection to configure script
This commit adds a --with-ofi-provider=<prov_name> flag to allow the user to
specify a provider to statically optimize for. Currently, the only
optimization supported is defining the GASNETC_OFI_HAS_MR_SCALABLE branch
condition to either 0 or 1 (or undefined to enable runtime detection).

Additionally, this commit adds auto-detection of installed providers using the
fi_info utility. It does this by iterating through a list of supported providers
and calling fi_info --provider <prov name>.

One issue that can come up with this approach is that if, for example, the gni
provider one day supports FI_MR_SCALABLE, there will be a mismatch between
what the configure script sets and what the provider wants. To address this, a
--{enable,disable}-ofi-mr-scalable was added as well to override the default
value.