psm: prevent a race from disabling poll_periodic
Prior to this commit, a race between two (or more) threads performing
the increment in gasnetc_psm_poll_periodic() could have resulted in it
reaching a value of 33 or higher and the periodic polling would cease
until the value had wrapped around back to 32.

This commit changes the condition to ">= 32".  Now the same race may
result in multiple threads polling simultaneously, but that is
harmless and prefered to *eliminating* the race using an atomic
fetch-add, mutex or similar.

NOTE: by the C specification the result of the race is probably
totally undefined, but we are making the reasonable assuption that
only "sane" implementaions (load, add, store) are in use.