ofi: Fix byte count in bounce buffer rma path
This commit fixes a typo that was made when adding FI_MR_BASIC support to ofi
conduit. This typo was causing the transfer of data outside of the limits of a
bounce buffer to take place.

This is expected to fix bug 3416.

(cherry picked from commit a7279283f5a8d45c68d78f1e61b20acb3b4b2679)