psm: detect bug 3419, report and abort
This commit ensures that in the case of unequal values for psm2's
max_request_short the job will abort at the end of gasnet_attach().
The resulting error message directs the user to the README, which in
turn directs the reader to the corresponding bugzilla entry where one
can expect to find the most current information.