ofi: Locking changes to fix bug 3417
This commit adds/edits locking macros in the ofi conduit in order to address
bug 3417 that is causing heap corruption with the psm2 provider. The
configure script adds a new global macro GASNETC_OFI_USE_THREAD_DOMAIN to
determine which threading mode to use, which is set to FI_THREAD_DOMAIN if
using psm/psm2 and otherwise FI_THREAD_FID.

In SEQ and PARSYNC  mode, the locking macros compile to empty do-while
loops, with the exception of GASNETC_OFI_TRYLOCK which compiles to 0 to
imitate successfully acquiring a lock.

In PAR mode, if GASNETC_OFI_USE_THREAD_DOMAIN is non-zero, all lock calls
will compile down to using a single "big" lock. The calls still look like
they are taking a fine-grained lock as an argument, but the fine grain lock
is ignored in the macro. If GASNETC_OFI_USE_THREAD_DOMAIN is zero, the lock
calls will compile to actually using the fine grained locks.