More cross-configure-bgq updates for bgclang
This commit removes "-fno-integrated-as" from the definitions of CC
and CXX for the (ALCF specific) bgclang environment for BG/Q.  This
flag was previously added to work around a bug in the integrated
assembler which has since been fixed.  With the current version of
bgclang this flag was causing an error rather than avoiding one!