pmi-spawner: fix another problem when MPI absent
This commit ensures that when performing pmi-based spawning the
gasnetrun_[conduit] perl script will look for "GASNetPMISpawner: 1"
when attempting to validate the executable.  Prior to this commit
"MPI" was substututed for "PMI", which was fine as long as both were
compiled.  However, in the absence of mpi-spawner support, this lead
to a warning:
  gasnetrun: unable to locate a GASNet program in [full command]

That, in turn, indicated another problem since the expected outcome
of simply looking for the wrong pattern was a fatal error:
  GASNet executable '[excutable]' does not support spawner 'MPI'
So, this commit also adds a missing backslash in a perl regex to
correct that misbehavior.

Note: it was because this second bug replaced the fatal error with a
warning that I failed to notice the first bug until now.  Sigh.