bug3430: workaround for pthread_create overcommit on cygwin
Add a workaround to prevent the bad performance interaction
observed on cygwin with pthread_create() and concurrent spin
calls to sched_yield on all cores.

The workaround causes sched_yield calls made concurrently
with a pthread_create call to sleep on a condition variable,
thus forcing them to yield the processor.