Configure: deal better w/ unified /bin /usr/bin
Some Linux distributions have been moving toward unifying /bin and
/usr/bin (by making one a symlink to the other).  As a result, the
configure probes which makes guesses based on an executable (ibv, ofi
and mxm) have begun to find their respective executables in /bin and
as a result they add "-I//include" to the conduit-specific includes.

This commit expands recently added logic to special-case "/usr" to
also check for "/", and switches from "test" to "case" for the
matching.  This includes making this change for psm-conduit, which is
not currently using an executable-based guess.